services/ingestion_service/requirements.txt
requests==2.31.0
sqlalchemy==2.0.25
psycopg[binary]==3.1.18
python-dotenv==1.0.1
tenacity==8.2.3